+++ linux/flutter/generated_plugins.cmake
@@ -5,6 +5,7 @@
 list(APPEND FLUTTER_PLUGIN_LIST
   desktop_drop
   dynamic_color
+  file_saver
   file_selector_linux
   flutter_timezone
   objectbox_flutter_libs